+++ src/snowflake/snowpark/modin/plugin/_internal/indexing_utils.py
@@ -478,19 +478,40 @@ def get_frame_by_row_pos_slice_frame(
     # Row position column required for left and right bound comparison.
     frame = internal_frame.ensure_row_position_column()
     row_pos_col = col(frame.row_position_snowflake_quoted_identifier)
+    count_col_created = False
 
-    # create a count column which will be used to check step
-    frame = frame.append_column("count", pandas_lit(1) + max_(row_pos_col).over())
-    count_col = col(frame.data_column_snowflake_quoted_identifiers[-1])
+    def get_count_col() -> Column:
+        nonlocal count_col_created
+        nonlocal frame
+        if not count_col_created:
+            # create a count column which will be used to check step
+            frame = frame.append_column(
+                "count", pandas_lit(1) + max_(row_pos_col).over()
+            )
+            count_col_created = True
+        return col(frame.data_column_snowflake_quoted_identifiers[-1])
 
     ordering_columns = internal_frame.ordering_columns
     start, stop, step = key.start, key.stop, key.step
 
     def make_positive(val: int) -> Column:
         # Helper to turn negative start and stop values to positive values. Example: -1 --> num_rows - 1.
-        return val + count_col if val < 0 else pandas_lit(val)
+        return val + get_count_col() if val < 0 else pandas_lit(val)
 
     step = 1 if step is None else step
+
+    # if `limit_n` is not None, use limit in the query to narrow down the search.
+    limit_n = None
+    if step < 0:
+        # Switch start and stop; convert given slice key into a similar slice key with positive step.
+        start, stop = stop, start
+    if (stop is not None and stop >= 0) and (start is None or start >= 0):
+        limit_n = stop + 1 if step < 0 else stop
+        if start is not None:
+            limit_n = limit_n - start
+            if limit_n < 0:
+                limit_n = 0
+
     if step < 0:
         # Set ascending to False if step is negative.
         ordering_columns = [
@@ -498,14 +519,12 @@ def get_frame_by_row_pos_slice_frame(
                 internal_frame.row_position_snowflake_quoted_identifier, ascending=False
             )
         ]
-        # Switch start and stop; convert given slice key into a similar slice key with positive step.
-        start, stop = stop, start
         start = 0 if start is None else make_positive(start) + 1
-        stop = count_col - 1 if stop is None else make_positive(stop)
+        stop = get_count_col() - 1 if stop is None else make_positive(stop)
     else:  # step > 0
         # Assign default values or convert to positive values.
         start = pandas_lit(0) if start is None else make_positive(start)
-        stop = (count_col if stop is None else make_positive(stop)) - 1
+        stop = (get_count_col() if stop is None else make_positive(stop)) - 1
 
     # Both start and stop are inclusive.
     left_bound_filter = row_pos_col >= start
@@ -514,13 +533,13 @@ def get_frame_by_row_pos_slice_frame(
     if step > 1:
         # start can be negative --> make the lower-bound 0.
         step_bound_filter = (
-            (row_pos_col - greatest(pandas_lit(0), least(start, count_col - 1)))
+            (row_pos_col - greatest(pandas_lit(0), least(start, get_count_col() - 1)))
             % pandas_lit(step)
         ) == 0
     elif step < -1:
         # Similarly, stop can be too large --> make the upper-bound the max row number.
         step_bound_filter = (
-            (greatest(pandas_lit(0), least(stop, count_col - 1)) - row_pos_col)
+            (greatest(pandas_lit(0), least(stop, get_count_col() - 1)) - row_pos_col)
             % pandas_lit(abs(step))
         ) == 0
     else:  # abs(step) == 1, so all values in range are included.
@@ -528,6 +547,8 @@ def get_frame_by_row_pos_slice_frame(
 
     filter_cond = left_bound_filter & right_bound_filter & step_bound_filter
     ordered_dataframe = frame.ordered_dataframe.filter(filter_cond)
+    if limit_n is not None:
+        ordered_dataframe = ordered_dataframe.limit(limit_n, sort=False)
     ordered_dataframe = ordered_dataframe.sort(ordering_columns)
     return InternalFrame.create(
         ordered_dataframe=ordered_dataframe,